library/util_axis_fifo_asym: fix tlast behavior when tkeep=0
Fix tvalid low when tlast is present.

Before this commit, util_axis_fifo would suppress all transfers with tkeep=0,
even if tlast was asserted. This commit makes it so transfers with tlast
asserted are preserved.

The AXI-Streaming spec allows us to suppress transfers for which all tkeep bits
are 0 (all bytes are null bytes), but only in the case when tlast=0 as well.
Transfers where tlast is asserted can't be suppressed even when all bytes are
null.

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>