projects/adrv9009zu11eg: Corundum Integration
- Created a new constraints file, alongside with
a new top file, for routing the QSFP signals
- Added a separate bd.tcl script, to handle the
necessary exports, configurations and connections
when building to support Corundum features
- common/adrv9009zu11eg_bd.tcl: DMA clock configuration
has a different value (250 MHz) if it's build for Corundum
integration. This value is exported as and env. var. from Make
and selected in the tcl scprit. Default it's set to (332.9 MHz)

Signed-off-by: Cristian Mihai Popa <cristianmihai.popa@analog.com>